refactor(Menu): DisplayName 적용
